+++ tests/oauth/test_credentials_flow.py
@@ -1,46 +1,11 @@
-import base64
-import json
 import time
-
-import pytest
-import requests
-import responses
+import base64
 from unittest.mock import patch
 
+import pytest
+import responses
+
 import sptfy.oauth as oauth
-
-
-CRED_INIT_VARS = [
-        (
-            {'client_id': 'client-id-123456', 'client_secret': 'secret-key-456789'},
-            {'scope': [], 'state':None, 'redirect_uri': None}
-        ),
-        (
-            {'client_id': 'client-id-123456', 'client_secret':'secret-key-456789', 'redirect_uri': 'http://localhost:4200/'},
-            {'scope': [], 'state': None}
-        ),
-        (
-            {'client_id': 'client-id-123456', 'client_secret': 'secret-key-456789', 'redirect_uri': 'http://localhost:4200/', 'state': 'random-state-string'},
-            {'scope': []}
-        ),
-        (
-            {'client_id': 'client-id-123456', 'client_secret': 'secret-key-456789', 'redirect_uri': 'http://localhost:4200/', 'state': 'random-state-string', 'scope': ['foo', 'bar']},
-            {},
-        )
-]
-
-
-@pytest.fixture()
-def sptfy_environment(monkeypatch):
-    client_id = 'test-client-id-123456'
-    client_secret = 'test-secret-key-456789'
-    redirect_uri = 'http://localhost:4200/'
-
-    monkeypatch.setenv('SPTFY_CLIENT_ID', client_id)
-    monkeypatch.setenv('SPTFY_CLIENT_SECRET', client_secret)
-    monkeypatch.setenv('SPTFY_REDIRECT_URI', redirect_uri)
-
-    return client_id, client_secret, redirect_uri
 
 
 def authorization_header(client_id, client_secret):
@@ -49,124 +14,18 @@ def authorization_header(client_id, client_secret):
     return {'Authorization': f"Basic {authorization_header.decode('ascii')}"}
 
 
-def test_load_credentials_from_env(sptfy_environment):
-    # GIVEN: The necessary environment variables
-    client_id, client_secret, redirect_uri = sptfy_environment
+unmocked = time.time
+def time_initial_values(values=None):
+    if values is None:
+        values = []
+    iter_val = iter(values)
 
-    # WHEN: ClientCredentials is created from env variables
-    credentials = oauth.ClientCredentials.from_env_variables()
-
-    # THEN: We should be able to access the client identification
-    assert credentials.client_id == client_id
-    assert credentials.client_secret == client_secret
-    assert credentials.redirect_uri == redirect_uri
-
-
-def test_load_credentials_without_env_should_throw():
-    # GIVEN: No environment variables
-    # WHEN: Initialized with from_env_variables()
-    # THEN: It should raise an exception
-    with pytest.raises(oauth.SptfyOAuthError) as err:
-        oauth.ClientCredentials.from_env_variables()
-
-    assert "SPTFY_CLIENT_ID" in str(err) or "SPTFY_CLIENT_SECRET" in str(err)
-
-
-@pytest.mark.parametrize("test_input,defaults", CRED_INIT_VARS)
-def test_credentials_initialization(test_input, defaults):
-
-    # WHEN: Initialized with custom variables
-    credentials = oauth.ClientCredentials(**test_input)
-
-    # THEN: Credentials should be initialized with default values
-    for key, value in defaults.items():
-        assert getattr(credentials, key) == value
-
-
-def test_credentials_initialization_has_default_variables(sptfy_environment):
-    # GIVEN: The credentials stored in program variables
-    client_id, client_secret, redirect_uri = sptfy_environment
-
-    # WHEN: Initialized only with client ID and client secret
-    credentials = oauth.ClientCredentials(client_id, client_secret)
-
-    # THEN: It should have an empty scope and no state
-    assert credentials.redirect_uri is None
-    assert credentials.scope == []
-    assert credentials.state is None
-
-
-def test_credentials_get_auth_header(sptfy_environment):
-    # GIVEN: A client id and a client_secret
-    client_id, client_secret, _ = sptfy_environment
-    credentials = oauth.ClientCredentials(client_id, client_secret)
-    auth_header = authorization_header(credentials.client_id, credentials.client_secret)
-    
-    # WHEN: authentication_header is called
-    header = credentials._authorization_header()
-
-    # THEN: The header should be a dict containing a base64 encoded
-    # string that represents the client_id and client_secret
-    assert header['Authorization'] == auth_header['Authorization']
-
-
-def test_oauth_token_from_json():
-    # GIVEN: A dict with the token information
-    token = {
-        'access_token':  'random-string-from-server',
-        'scope': 'foo bar',
-        'expires_in': 3600,
-        'token_type': 'Bearer',
-    }
-
-    # WHEN: loading an OAuthToken from a json dict
-    oauth_token = oauth.OAuthToken.from_json(token)
-
-    # THEN: the scope parsed into a list
-    assert oauth_token.scope == ['foo', 'bar']
-    # Refresh token is optional (client credentials flow doesnt use it)
-    assert oauth_token.refresh_token == None
-
-
-@patch('sptfy.oauth.time.time')
-def test_oauth_token_expires_at(mock_time):
-    # GIVEN: an expiration timespan from the token
-    token = {
-        'access_token':  'random-string-from-server',
-        'scope': 'foo bar',
-        'expires_in': 3600,
-        'token_type': 'Bearer',
-    }
-
-    mock_time.return_value = 10 
-    oauth_token = oauth.OAuthToken.from_json(token)
-
-    # WHEN: expires_at is called 
-    # THEN: The expiration timestamp should be returned
-    assert oauth_token.expires_at == mock_time.return_value + token['expires_in']
-    oauth_token.expires_at
-    # the expiration timestamp should not be recalculated
-    assert mock_time.call_count == 1
-
-
-@patch('sptfy.oauth.time.time')
-def test_oauth_token_is_expired(mock_time):
-    # GIVEN: An expired token
-    mock_time.side_effect = [0, 30, 101]
-
-    json_token = {
-        'access_token':  'random-string-from-server',
-        'scope': 'foo bar',
-        'expires_in': 100,
-        'token_type': 'Bearer',
-    }
-
-    token = oauth.OAuthToken.from_json(json_token)
-
-    # THEN: token should not be expired after 30 milliseconds (i think, maybe seconds)
-    assert not token.is_expired
-    # token should be expired after 100 milliseconds
-    assert token.is_expired
+    def time_side_effect():
+        try:
+            return next(iter_val)
+        except StopIteration:
+            return unmocked()
+    return time_side_effect
 
 
 def test_oauth_client_credentials_flow_from_env(sptfy_environment):
@@ -268,18 +127,7 @@ def test_oauth_client_credentials_successful_get_access_token(sptfy_environment)
     assert credentials_flow._current_token == token
 
 
-unmocked = time.time
-def time_initial_values(values=None):
-    if values is None:
-        values = []
-    iter_val = iter(values)
 
-    def time_side_effect():
-        try:
-            return next(iter_val)
-        except StopIteration:
-            return unmocked() 
-    return time_side_effect
 
 
 @responses.activate
